# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gomeshcomd API
  version: 0.10.0
  description: |
    HTTP and Server-Sent Events API exposed by gomeshcomd.
    Static frontend routes are served by the same HTTP listener but are not
    part of this API contract.
servers:
  - url: http://127.0.0.1:8080
    description: Default local server
...
security:
  - sessionCookie: []
paths:
...
  /api/config:
    get:
      tags: [Config]
      summary: Return the current daemon configuration
      description: |
        Returns the full effective configuration including which fields are
        currently overridden by environment variables and which require a restart
        to take effect. The `auth.password` value is always masked as `"****"`.

        Returns `403` when demo mode is active.
      operationId: getConfig
      responses:
        "200":
          description: Current effective configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Config API disabled in demo mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Config]
      summary: Partially update the daemon configuration
      description: |
        Merges the supplied fields onto the current effective configuration.
        Omitted fields are left unchanged. Duration values use Go duration
        strings (e.g. `"40s"`, `"2h"`); storage purge durations also accept
        day suffixes such as `"30d"`.

        Fields currently managed by environment variables cannot be changed via
        this endpoint and return `409`. Fields that require a daemon restart are
        flagged in the response via `requires_restart: true`.

        On success the updated configuration is written to the TOML config file
        (if `WithTomlPath` is configured); a changed callsign also triggers an
        immediate `station.identity` SSE broadcast.

        Returns `403` when demo mode is active.
      operationId: updateConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdateRequest"
            examples:
              changeLogLevel:
                value:
                  log_level: debug
              changeChatWindow:
                value:
                  chat_log:
                    history_window: "48h"
      responses:
        "200":
          description: Updated configuration and restart flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigUpdateResponse"
        "400":
          description: Invalid JSON or invalid field value (e.g. bad duration string).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Config API disabled in demo mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Field is locked by an environment variable and cannot be changed at runtime.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: TOML persistence failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid json
...
    ConfigFieldMeta:
      type: object
      required: [value, env_override, requires_restart]
      description: A single configuration field with its metadata.
      properties:
        value:
          description: Current effective value of the field.
        env_override:
          type: boolean
          description: True when this field is currently managed by an environment variable and cannot be changed via the API.
          example: false
        requires_restart:
          type: boolean
          description: True when changing this field requires a daemon restart to take effect.
          example: true
    ConfigServerInfo:
      type: object
      required: [version, started_at, uptime_seconds]
      properties:
        version:
          type: string
          description: Daemon version string.
          example: v0.9.0
        started_at:
          type: string
          format: date-time
          description: Time the daemon started (UTC).
          example: "2026-06-13T08:00:00Z"
        uptime_seconds:
          type: integer
          format: int64
          description: Seconds since daemon start.
          example: 7200
    ConfigReceiveLogSection:
      type: object
      required: [enabled, path, retention_days, replay_window]
      properties:
        enabled:
          $ref: "#/components/schemas/ConfigFieldMeta"
        path:
          $ref: "#/components/schemas/ConfigFieldMeta"
        retention_days:
          $ref: "#/components/schemas/ConfigFieldMeta"
        replay_window:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigStatsSection:
      type: object
      required: [enabled, path, retention_days]
      properties:
        enabled:
          $ref: "#/components/schemas/ConfigFieldMeta"
        path:
          $ref: "#/components/schemas/ConfigFieldMeta"
        retention_days:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigChatLogSection:
      type: object
      required: [path, history_window, max_history_window]
      properties:
        path:
          $ref: "#/components/schemas/ConfigFieldMeta"
        history_window:
          $ref: "#/components/schemas/ConfigFieldMeta"
        max_history_window:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigSendSection:
      type: object
      required: [dedup_ttl]
      properties:
        dedup_ttl:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigForwardSection:
      type: object
      required: [targets]
      properties:
        targets:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigAuthSection:
      type: object
      required: [username, password, session_ttl, cookie_name]
      description: Authentication configuration. `password.value` is always `"****"`.
      properties:
        username:
          $ref: "#/components/schemas/ConfigFieldMeta"
        password:
          $ref: "#/components/schemas/ConfigFieldMeta"
        session_ttl:
          $ref: "#/components/schemas/ConfigFieldMeta"
        cookie_name:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigRequestLogSection:
      type: object
      required: [enabled]
      properties:
        enabled:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigStorageSection:
      type: object
      required: [sqlite_path, purge_interval, receive_log_retention, public_chat_retention, nodes_retention, telemetry_retention]
      properties:
        sqlite_path:
          $ref: "#/components/schemas/ConfigFieldMeta"
        purge_interval:
          $ref: "#/components/schemas/ConfigFieldMeta"
        receive_log_retention:
          $ref: "#/components/schemas/ConfigFieldMeta"
        public_chat_retention:
          $ref: "#/components/schemas/ConfigFieldMeta"
        nodes_retention:
          $ref: "#/components/schemas/ConfigFieldMeta"
        telemetry_retention:
          $ref: "#/components/schemas/ConfigFieldMeta"
    ConfigResponse:
      type: object
      required:
        - server
        - http_addr
        - udp_listen_addr
        - node_addr
        - my_call
        - data_dir
        - max_message_length
        - log_level
        - receive_log
        - stats
        - chat_log
        - send
        - forward
        - auth
        - request_log
        - storage
      properties:
        server:
          $ref: "#/components/schemas/ConfigServerInfo"
        http_addr:
          $ref: "#/components/schemas/ConfigFieldMeta"
        udp_listen_addr:
          $ref: "#/components/schemas/ConfigFieldMeta"
        node_addr:
          $ref: "#/components/schemas/ConfigFieldMeta"
        my_call:
          $ref: "#/components/schemas/ConfigFieldMeta"
        data_dir:
          $ref: "#/components/schemas/ConfigFieldMeta"
        max_message_length:
          $ref: "#/components/schemas/ConfigFieldMeta"
        log_level:
          $ref: "#/components/schemas/ConfigFieldMeta"
        receive_log:
          $ref: "#/components/schemas/ConfigReceiveLogSection"
        stats:
          $ref: "#/components/schemas/ConfigStatsSection"
        chat_log:
          $ref: "#/components/schemas/ConfigChatLogSection"
        send:
          $ref: "#/components/schemas/ConfigSendSection"
        forward:
          $ref: "#/components/schemas/ConfigForwardSection"
        auth:
          $ref: "#/components/schemas/ConfigAuthSection"
        request_log:
          $ref: "#/components/schemas/ConfigRequestLogSection"
        storage:
          $ref: "#/components/schemas/ConfigStorageSection"
    ConfigUpdateReceiveLog:
      type: object
      properties:
        enabled:
          type: boolean
        path:
          type: string
        retention_days:
          type: integer
        replay_window:
          type: string
          description: Go duration string, e.g. `"2h"`.
          example: "2h"
    ConfigUpdateStats:
      type: object
      properties:
        enabled:
          type: boolean
        path:
          type: string
        retention_days:
          type: integer
    ConfigUpdateChatLog:
      type: object
      properties:
        path:
          type: string
        history_window:
          type: string
          description: Go duration string, e.g. `"48h"`.
          example: "48h"
        max_history_window:
          type: string
          description: Go duration string, e.g. `"720h"`.
          example: "720h"
    ConfigUpdateSend:
      type: object
      properties:
        dedup_ttl:
          type: string
          description: Go duration string, e.g. `"30s"`.
          example: "30s"
    ConfigUpdateForward:
      type: object
      properties:
        targets:
          type: string
          description: Comma-separated list of UDP forward targets (host:port).
          example: "192.168.0.5:1799,10.0.0.1:1799"
    ConfigUpdateAuth:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        session_ttl:
          type: string
          description: Go duration string, e.g. `"24h"`.
          example: "24h"
        cookie_name:
          type: string
          example: meshcom_session
    ConfigUpdateRequestLog:
      type: object
      properties:
        enabled:
          type: boolean
    ConfigUpdateStorage:
      type: object
      properties:
        sqlite_path:
          type: string
          example: "./data/gomeshcom.db"
        purge_interval:
          type: string
          description: Duration string. TOML/API accept day suffixes such as `"30d"` in addition to Go duration strings.
          example: "4h"
        receive_log_retention:
          type: string
          description: Duration string. TOML/API accept day suffixes such as `"30d"` in addition to Go duration strings.
          example: "30d"
        public_chat_retention:
          type: string
          description: Duration string. TOML/API accept day suffixes such as `"30d"` in addition to Go duration strings.
          example: "30d"
        nodes_retention:
          type: string
          description: Duration string. TOML/API accept day suffixes such as `"7d"` in addition to Go duration strings.
          example: "7d"
        telemetry_retention:
          type: string
          description: Duration string. TOML/API accept day suffixes such as `"30d"` in addition to Go duration strings.
          example: "30d"
    ConfigUpdateRequest:
      type: object
      description: |
        Partial configuration update. Omitted fields are left unchanged.
        Duration values are Go duration strings (e.g. `"40s"`, `"2h"`).
        Storage purge durations also accept day suffixes such as `"30d"`.
        Fields managed by environment variables return `409` if present.
      properties:
        http_addr:
          type: string
          example: "0.0.0.0:8080"
        udp_listen_addr:
          type: string
          example: "0.0.0.0:1799"
        node_addr:
          type: string
          example: "192.168.0.2:1799"
        my_call:
          type: string
          example: IU5PMP-1
        max_message_length:
          type: integer
          example: 160
        log_level:
          type: string
          enum: [debug, info, warn, error]
          example: info
        receive_log:
          $ref: "#/components/schemas/ConfigUpdateReceiveLog"
        stats:
          $ref: "#/components/schemas/ConfigUpdateStats"
        chat_log:
          $ref: "#/components/schemas/ConfigUpdateChatLog"
        send:
          $ref: "#/components/schemas/ConfigUpdateSend"
        forward:
          $ref: "#/components/schemas/ConfigUpdateForward"
        auth:
          $ref: "#/components/schemas/ConfigUpdateAuth"
        request_log:
          $ref: "#/components/schemas/ConfigUpdateRequestLog"
        storage:
          $ref: "#/components/schemas/ConfigUpdateStorage"
    ConfigUpdateResponse:
      type: object
      required: [config, requires_restart]
      properties:
        config:
          $ref: "#/components/schemas/ConfigResponse"
        requires_restart:
          type: boolean
          description: True when at least one changed field requires a daemon restart to take effect.
          example: false
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: meshcom_session
      description: |
        HTTP-only session cookie created by `POST /api/session`. Its actual
        name is configured by `GOMESHCOM_AUTH_COOKIE_NAME`.